+++ src/pages/member/GetAuthMain.jsx
@@ -16,8 +16,8 @@ const GetAuthMain = () => {
 
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const loginError = useSelector((state) => state.auth.loginError);
-  const [isAuthInitialized, setIsAuthInitialized] = useState(false);
+  // const loginError = useSelector((state) => state.auth.loginError);
+  // const [isAuthInitialized, setIsAuthInitialized] = useState(false);
   
   const handleLoginClick = () => {
     console.log("$$$$ 일반 로그인 버튼 클릭");
@@ -27,58 +27,58 @@ const GetAuthMain = () => {
   const handleGoogleLoginClick = () => {
     console.log("$$$$ 구글 로그인 버튼 클릭");
     clearAllCookies();
-    handleGoogleLogin(dispatch);
+    handleGoogleLogin(dispatch,navigate);
   };
 
-  useEffect(() => {
-    // Firebase auth 객체가 초기화되는지 확인
-    const checkAuthInitialization = () => {
-      if (auth) {
-        // alert("Firebase auth 객체가 초기화되었습니다.");
-        setIsAuthInitialized(true);
-      } else {
-        //alert("Firebase auth 객체가 초기화되지 않았습니다.");
-        console.error("Firebase auth 객체가 초기화되지 않았습니다.");
-      }
-    };
+  // useEffect(() => {
+  //   // Firebase auth 객체가 초기화되는지 확인
+  //   const checkAuthInitialization = () => {
+  //     if (auth) {
+  //       // alert("Firebase auth 객체가 초기화되었습니다.");
+  //       setIsAuthInitialized(true);
+  //     } else {
+  //       //alert("Firebase auth 객체가 초기화되지 않았습니다.");
+  //       console.error("Firebase auth 객체가 초기화되지 않았습니다.");
+  //     }
+  //   };
 
-    checkAuthInitialization();
-  }, []);
+  //   checkAuthInitialization();
+  // }, []);
 
 
-  useEffect(() => {
-    if (!isAuthInitialized) {
-      return;
-    }
+  // useEffect(() => {
+  //   if (!isAuthInitialized) {
+  //     return;
+  //   }
 
-    console.log("@@@@ GetAuthMain useEffect - checkRedirectResult 호출");
-    console.log(`loginError`, loginError);
+  //   console.log("@@@@ GetAuthMain useEffect - checkRedirectResult 호출");
+  //   console.log(`loginError`, loginError);
     
-    const getResult = async () => {
-      console.log("$$$$ getResult 함수 시작");
-      const response = await checkRedirectResult(dispatch);
-      console.log("checkRedirectResult 결과:", response); // 여기는 토큰과 유저가 온다.
-      // alert("checkRedirectResult 결과:"+ response); // 여기는 토큰과 유저가 온다.
-      const info = response?.user;
-      const token = response?.token;
-      if (info?.email && info?.email !== "") {
-        const loginData = {
-          member: {
-            authorizeCode: token,
-            memberId: info?.email,
-            nickname: info?.displayName,
-            profilePhotoUrl: info?.photoURL,
-            phoneNumber: info?.phoneNumber
-          },
-          loginWay: "google"
-        };
-        console.log("$$$$ 구글 로그인 시도");
-        dispatch(login({ loginData, navigate }));
-      }
-    };
+  //   const getResult = async () => {
+  //     console.log("$$$$ getResult 함수 시작");
+  //     const response = await checkRedirectResult(dispatch);
+  //     console.log("checkRedirectResult 결과:", response); // 여기는 토큰과 유저가 온다.
+  //     // alert("checkRedirectResult 결과:"+ response); // 여기는 토큰과 유저가 온다.
+  //     const info = response?.user;
+  //     const token = response?.token;
+  //     if (info?.email && info?.email !== "") {
+  //       const loginData = {
+  //         member: {
+  //           authorizeCode: token,
+  //           memberId: info?.email,
+  //           nickname: info?.displayName,
+  //           profilePhotoUrl: info?.photoURL,
+  //           phoneNumber: info?.phoneNumber
+  //         },
+  //         loginWay: "google"
+  //       };
+  //       console.log("$$$$ 구글 로그인 시도");
+  //       dispatch(login({ loginData, navigate }));
+  //     }
+  //   };
 
-    getResult();
-  }, [dispatch, loginError, navigate, isAuthInitialized]);
+  //   getResult();
+  // }, [dispatch, loginError, navigate, isAuthInitialized]);
 
   return (
     <ForegroundTemplate>
